tests/wpt/web-platform-tests/old-tests/submission/Infraware/Forms/contents/Forms/form_length.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>Form_length</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      id="input_form">
  <p><input type=hidden name="custname"></p>
  <p><input type=hidden name="custtel"></p>
  <p><input type=hidden name="custemail"></p>
 
  </form>
  <script>

    var form = document.getElementById("input_form");
    var len = form.length;

    test(function() {
      assert_equals(len, 3, "length attribute is not correct.");
    });

  </script>

 </body>
</html>
